Refactor to fit project structure
