+++ packages/harmony/src/components/hover-card/index.ts
@@ -1,4 +1,6 @@
 export { FlairHoverCard } from './FlairHoverCard'
-export type { FlairHoverCardProps } from './FlairHoverCard'
 export { BaseFlairHoverCardHeader } from './FlairHoverCardHeader'
-export type { BaseFlairHoverCardHeaderProps } from './FlairHoverCardHeader'
+export type {
+  FlairHoverCardProps,
+  BaseFlairHoverCardHeaderProps
+} from './types'
